improve SttpUtil error handling and consistently add retries.
